Refactoring Enemy Movement and Bullet Code
Refactoring Done
- Separated Movement scripts from Bullet scripts
Note: Movement and Bullet Scripts should be completely independent from each other, with the reason being that we should be able to mix and match movement and bullet scripts.

Further Refactoring needs to be done
- Change MoveLinearToPos to have a constant start velocity rather than updating position every physics frame.
- Make comments on every script for brief overviews of the code
- Change the names of some scripts to better articulate what the scripts do
- Make general efficiency and simplicity changes across all scripts.
