+++ Side Scroller Shooter/Assets/Prefabs/EnemyShooterBasic.prefab
@@ -45,6 +45,7 @@ GameObject:
   - component: {fileID: 8209959607371746374}
   - component: {fileID: 8209959607371746377}
   - component: {fileID: 873641348055334267}
+  - component: {fileID: -7463063941175694100}
   m_Layer: 0
   m_Name: EnemyShooterBasic
   m_TagString: Enemy
@@ -130,7 +131,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 1149adfa276b4914fbba38729d25fafd, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  speed: 3
   bullet: {fileID: 187162364329538102, guid: 6256d480f56c12547bfe898fd6f086a7, type: 3}
   bulletParent: {fileID: 8209959606965187190}
   fireRate: 1
@@ -207,3 +207,16 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   damage: 10
+--- !u!114 &-7463063941175694100
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8209959607371746371}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9c0cd87360c33504f9a21e09fc29a60a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  moveSpeed: 2
