+++ Side Scroller Shooter/Assets/Scripts/Enemy/Bullets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e7912177aca7e564ea9bf65c41166b99
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
