+++ Side Scroller Shooter/Assets/Scripts/Enemy/Bullets/BulletHoming.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class EnemyHomingBullet : MonoBehaviour
+public class BulletHoming : MonoBehaviour
 {
     [SerializeField] public float speed;
     Transform player;
