+++ Side Scroller Shooter/Assets/Scripts/Enemy/EnemyGun.cs
@@ -2,9 +2,8 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class EnemyFollow : MonoBehaviour
+public class EnemyGun : MonoBehaviour
 {
-    [SerializeField] private float speed;
     private Transform player;
     public GameObject bullet;
     public GameObject bulletParent;
@@ -20,7 +19,6 @@ public class EnemyFollow : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        transform.position = Vector2.MoveTowards(this.transform.position, player.position, speed * Time.deltaTime);
         if (nextFireTime < Time.time)
         {
             Instantiate(bullet, bulletParent.transform.position, Quaternion.identity);
