--- Side Scroller Shooter/Assets/Scripts/Enemy/EnemyMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bfd2e1362eb9f814e806ebc0d6cc6a11
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
