+++ Side Scroller Shooter/Assets/Scripts/Enemy/Movement.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 30124c65d762fe643af452cff168250b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
