+++ Side Scroller Shooter/Assets/Scripts/Enemy/Movement/MoveSine.cs
@@ -2,14 +2,18 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class EnemyMovement : MonoBehaviour
+public class MoveSine : MonoBehaviour
 {
+    float sinCenterY;
+    [SerializeField] float amplitude;
+    [SerializeField] float frequency;
+    [SerializeField] bool inverted;
     [SerializeField] float moveSpeed = 5;
 
     // Start is called before the first frame update
     void Start()
     {
-        
+        sinCenterY = transform.position.y;
     }
 
     // Update is called once per frame
@@ -22,8 +26,16 @@ public class EnemyMovement : MonoBehaviour
     {
         Vector2 pos = transform.position;
 
-        pos.x -= moveSpeed * Time.deltaTime;
+        //y position calculation
+        float sin = Mathf.Sin(pos.x * frequency) * amplitude;
+        if (inverted)
+        {
+            sin *= -1;
+        }
+        pos.y = sinCenterY + sin;
 
+        //x position calculation
+        pos.x -= moveSpeed * Time.deltaTime;
         if (pos.x < -15)
         {
             Destroy(gameObject);
